refactor: extract isClient to browser
