+++ packages/utils/browser.ts
@@ -1,4 +1,6 @@
-import { isClient } from './types'
+import { isClient, isIOS } from '@vueuse/core'
 
 export const isFirefox = (): boolean =>
   isClient && /firefox/i.test(window.navigator.userAgent)
+
+export { isClient, isIOS }
