+++ packages/utils/dom/scroll.ts
@@ -1,4 +1,4 @@
-import { isClient } from '../types'
+import { isClient } from '../browser'
 import { getStyle } from './style'
 
 export const isScroll = (el: HTMLElement, isVertical?: boolean): boolean => {
