+++ packages/utils/types.ts
@@ -10,7 +10,6 @@ export {
   isPromise,
   isSymbol,
 } from '@vue/shared'
-export { isClient, isIOS } from '@vueuse/core'
 export { isVNode } from 'vue'
 
 export const isUndefined = (val: any): val is undefined => val === undefined
