+++ packages/utils/vue/global-node.ts
@@ -1,4 +1,4 @@
-import { isClient } from '../types'
+import { isClient } from '../browser'
 
 const globalNodes: HTMLElement[] = []
 let target: HTMLElement = !isClient ? (undefined as any) : document.body
